Workaround for API hole bug
Sometimes there's a user that gets deleted faster than the bot can add it to the database. This results in their ID being reserved so newer players won't get added into the database, crashing the bot when they play matches.

This patch attaches to the crash and:
- Adds a new player
- Calls the function that adds new players so that all players are in db
- Calls the function where it crashed again with same arguments

Hopefully this solves this annoying bug!